rename package files to make_bcs*.py
